Extract features to separate module
